requirements.txt
# Mazrica → Google Sheets 同期機能の依存パッケージ

# HTTP クライアント
requests>=2.31.0

# Google API クライアント
google-api-python-client>=2.111.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

